+++ components/organisms/cardList/index.ts
@@ -1,0 +1,1 @@
+export * from "./cardList";
