+++ components/organisms/pagination/pagination.tsx
@@ -1,4 +1,4 @@
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/button";
 
 type PaginationProp = {
   data: any;
